Update 2014-03-27-showmodaldialog.md
Added final paragraph
Note:  The dialog-polyfill example on noted on GitHub is NOT compatible with IE, as it relies on the 2016 ECMAScript standard "Promise", which is not a construct supported by IE:  https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/Promise

When saying that an example is cross-browser, how about actually testing it and including IE?  As much as you guys may hate it, it is what those that live in the land of Microsoft products (i.e. SharePoint) must deal with everyday.  Thanks.